Permet de créer la fin d'une évaluation
Converti la requête `get` "compétences_fortes" en `post`
"evaluation/:id/fin"
